Support for prepositional arguments in ensembles
Ensembles may now take arguments that are passed before their subcommands,
specified in the optional second argument of the ensemble command. Passing the
placeholder ".." instead indicates that the argument is to be provided after
the selector. Such arguments must not be followed by non-placeholder ones;
this should be an error, but is not in the current implementation.

Support for the [itself] selector was clumsy and has been dropped.